+++ src/observer.h
@@ -32,7 +32,7 @@ public:
 	
 	void AddObserver(Observer<T>* observer)
 	{
-		int observer_num = observers_.size();
+		size_t observer_num = observers_.size();
 		for (int i = 0; i < observer_num; ++i)
 		{
 			if (observers_[i] == observer)
